+++ NanoXLSX4j.Lib/src/main/java/ch/rabanti/nanoxlsx4j/Workbook.java
@@ -310,7 +310,7 @@ public class Workbook {
 	// ### M E T H O D S ###
 
 	/**
-	 * Adds a color value (HEX; 6-digit RGB or 8-digit RGBA) to the MRU list
+	 * Adds a color value (HEX; 6-digit RGB or 8-digit ARGB) to the MRU list
 	 *
 	 * @param color
 	 *            RGB code in hex format (either 6 characters, e.g. FF00AC or 8